Basic framework for landing page made
