+++ silverplatter.client/src/App.css
@@ -1,11 +1,14 @@
 ﻿#root {
-    max-width: 1280px;
-    margin: 0 auto;
-    padding: 2rem;
-    text-align: center;
+    height: 100vh;
 }
 
-th, td {
-    padding-left: 1rem;
-    padding-right: 1rem;
+#Scaffold * {
+    outline: 2px dotted red;
+}
+
+#Scaffold {
+    outline: 2px dotted yellow;
+    width: 100%;
+    max-height: 100vh;
+    overflow-y: hidden;
 }
