+++ silverplatter.client/src/Components/Header.tsx
@@ -1,0 +1,15 @@
+
+
+function Header() {
+    return (
+        <div style={{
+            width: "100vw",
+            height: "5rem",
+            backgroundColor: "red"
+        }}>
+            THIS IS THE HEADER
+        </div>
+    )
+}
+
+export default Header;
