+++ silverplatter.client/src/Components/LandingPage.tsx
@@ -1,0 +1,28 @@
+
+
+function LandingPage() {
+
+
+    return (
+        <div style={{
+            width: "100vw",
+            backgroundColor: "black"
+        }}>
+            Landing Page
+
+            <div id="Content" style={{
+                marginTop: "1rem",
+                height: "100vh",
+                backgroundColor: "white",
+                color: "black"
+            }}>
+                This is the content
+                <img 
+                    src="silverplatter.client\src\assets\claim-your-i-was-here-button-v0-9tlo8368wdkf1.webp"
+                />
+            </div>
+        </div>
+    )
+}
+
+export default LandingPage;
